* 'develop' of https://gitlab.com/jusfoun-bigdata/hookah:
  bug
  邮件下发地址更换及代码完善
  fix
  fix bug
  结算处理
  下拉样式
  邮件下发地址更换及代码完善

+++ hookah-pay/src/main/java/com.jusfoun.hookah.pay/service/impl/PayAccountRecordServiceImpl.java
@@ -415,7 +415,7 @@ public class PayAccountRecordServiceImpl extends GenericServiceImpl<PayAccountRe
 		payAccountRecord.setUserId(moneyInOutBo.getUserId());		//当前操作用户的userID
 		payAccountRecord.setTransferDate(new Date());
 		payAccountRecord.setMoney(moneyInOutBo.getMoney());		//当前用户的入金出金的金额
-		payAccountRecord.setTransferType(moneyInOutBo.getOperatorType());	//根据操作类型  入金
+		payAccountRecord.setTransferType((byte) moneyInOutBo.getOperatorType().intValue());	//根据操作类型  入金
 		payAccountRecord.setTransferStatus(PayConstants.TransferStatus.handing.code);
 		payAccountRecord.setSerialNumber(serialNum);
 		payAccountRecord.setChannelType(ChannelType.QDABC);
